cleanup and feature tweaks
Brought code to osx 10.9.5 compatibility, such as change in gl headers,
or xcode build architecture setting.

CUSTOM render mode is now default for struct data.

fixed torsion angle calculation.

changed handling of altloc atoms; previously the altloc modifier would
be prepended to the resname, leading to confusion and crashes in any
sort of trace. Now altloc is ignored, the first loc is loaded as the
only one, subsequent atoms for that loc are ignored (not fully tested!)

removed runtime dependencies on libpng, libjpg, libtiff and libz as
these are no longer reliably present on OSX as a shared library. Only
-lm is used during linking.

some apparent changes are due to an accidental indent fix.